Change starting time of SingSparrow for the summer
The default time for automatic start-up of SingSparrow has been changed from 8am to 9am. This was done because birds in the Maney Lab are not moved into summer time, therefore, during summer time, the start-up time must be delayed by one hour. A patch is on the works to automatically change the start-up time when switching to and from summer time.