Add files for testing speaker output during experiment
New files were added to SingSparrow that can be used to test speaker output inside sound booth that already contain a bird. The files have simple tones, and therefore are not expected to be relevant to the bird. At this moment, the user has to manually start playback of each file; there is no utility in SingSparrow for playing them back.